# backend/tests/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strom API",
    "description": "REST and WebSocket API for the Strom pipeline engine",
    "contact": {
      "name": "Per Enstedt",
      "email": "per.enstedt@eyevinn.se"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.3.26"
  },
  "paths": {
...
    "/whep/{endpoint_id}": {
      "post": {
        "tags": [
          "whep"
        ],
        "summary": "Proxy POST requests to /whep/{endpoint_id}\nLooks up the internal port from WhepRegistry and forwards to localhost:{port}/whep/endpoint",
        "operationId": "whep_endpoint_proxy",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHEP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "WHEP session created, SDP answer returned",
            "content": {
              "application/sdp": {}
            }
          },
          "404": {
            "description": "WHEP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHEP server"
          }
        }
      },
      "options": {
        "tags": [
          "whep"
        ],
        "summary": "Handle OPTIONS preflight for /whep/{endpoint_id}",
        "operationId": "whep_endpoint_proxy_options",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHEP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight response"
          }
        }
      }
    },
    "/whep/{endpoint_id}/resource/{resource_id}": {
      "delete": {
        "tags": [
          "whep"
        ],
        "summary": "Proxy DELETE requests to /whep/{endpoint_id}/resource/{resource_id}",
        "operationId": "whep_resource_proxy_delete",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHEP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHEP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WHEP session deleted"
          },
          "404": {
            "description": "WHEP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHEP server"
          }
        }
      },
      "options": {
        "tags": [
          "whep"
        ],
        "summary": "Handle OPTIONS preflight for /whep/{endpoint_id}/resource/{resource_id}",
        "operationId": "whep_resource_proxy_options",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHEP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHEP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight response"
          }
        }
      },
      "patch": {
        "tags": [
          "whep"
        ],
        "summary": "Proxy PATCH requests to /whep/{endpoint_id}/resource/{resource_id} for ICE candidates",
        "operationId": "whep_resource_proxy_patch",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHEP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHEP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "ICE candidates accepted"
          },
          "404": {
            "description": "WHEP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHEP server"
          }
        }
      }
    },
...
  },
...
}